Fix device

+++ university-project/back/src/device/schemas.py
@@ -6,9 +6,9 @@ from pydantic import UUID4, BaseModel, EmailStr, constr, conint
 class DeviceSchema(BaseModel):
     id: Optional[str] = None
     name: str
-    AC_power_consumption: int
-    DC_power_consumption: int
+    AC_power_consumption: float
+    DC_power_consumption: float
 
 class DeviceUpdateSchema(BaseModel):
-    AC_power_consumption: int
-    DC_power_consumption: int
+    AC_power_consumption: float
+    DC_power_consumption: float
